+++ src/index.css
@@ -5,41 +5,3 @@
 .content-box {
   box-sizing: content-box;
 }
-
-.map-overlay {
-  width: 300px;
-  padding: 10px 20px 20px;
-  border-radius: 16px;
-  background: white;
-  color: black;
-}
-
-#close {
-  z-index: 10;
-  text-align: end;
-  cursor: pointer;
-}
-
-.name {
-  font-size: 24px;
-  font-weight: 600;
-}
-
-.desc {
-  margin-top: 6px;
-  text-wrap: wrap;
-  word-break: keep-all;
-  font-size: 14px;
-  color: #8f8f8f;
-}
-
-.phone-number {
-  margin-top: 5px;
-  color: #8f8f8f;
-  transition: all ease-in 0.2s;
-  font-size: 14px;
-}
-
-.phone-number:hover {
-  color: black;
-}
